+++ react-auth/src/TaskCard.tsx
@@ -19,12 +19,12 @@ export const TaskCard: FC<{ token: string }> = (props) => {
     };
 
     axios(configuration)
-        .then((result) => {
-          console.log(result.data.result.response.candidates[0].content.parts[0])
-        })
-        .catch((error) => {
-          console.log(error);
-        });
+      .then((result) => {
+        console.log(result.data.result.response.candidates[0].content.parts[0]);
+      })
+      .catch((error) => {
+        console.log(error);
+      });
   };
   const getTasks = () => {
     const configuration = {
@@ -139,91 +139,90 @@ export const TaskCard: FC<{ token: string }> = (props) => {
       <Card.Body>
         <Card.Title className="text-center">My Tasks</Card.Title>
 
-        <br/>
+        <br />
 
-        <Button onClick={getAi}>Gemini Test (check console devtools for output)</Button>
+        <Button onClick={getAi}>
+          Gemini Test (check console devtools for output)
+        </Button>
 
-        <br/>
-        <br/>
-
+        <br />
+        <br />
 
         <Form onSubmit={makeTask}>
           <Form.Group className="mb-3" controlId="formBasicEmail">
             <Form.Control
-                onChange={(e) => setTask(e.target.value)}
-                placeholder={
-                  taskList.length > 0 ? "Enter Task Name" : "Add your first task"
-                }
+              onChange={(e) => setTask(e.target.value)}
+              placeholder={
+                taskList.length > 0 ? "Enter Task Name" : "Add your first task"
+              }
             />
           </Form.Group>
         </Form>
 
         <Table>
           <tbody>
-          {taskList.map(
+            {taskList.map(
               (t) =>
-                  !t.completed && (
-                      <tr key={t._id}>
-                        <td className="align-middle">
-                          <input
-                              type="checkbox"
-                              checked={false}
-                              readOnly
-                              className="form-check-input"
-                              onClick={() => completeTasks(t.name)}
-                          />
-                        </td>
-                        <td className="align-middle">{t.name}</td>
-                        <td className="text-end align-middle">
-                          &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
-                        </td>
-                      </tr>
-                  ),
-          )}
+                !t.completed && (
+                  <tr key={t._id}>
+                    <td className="align-middle">
+                      <input
+                        type="checkbox"
+                        checked={false}
+                        readOnly
+                        className="form-check-input"
+                        onClick={() => completeTasks(t.name)}
+                      />
+                    </td>
+                    <td className="align-middle">{t.name}</td>
+                    <td className="text-end align-middle">
+                      &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
+                    </td>
+                  </tr>
+                ),
+            )}
 
-          {taskList.map(
+            {taskList.map(
               (t) =>
-                  t.completed && (
-                      <tr key={t._id}>
-                        <td className="align-middle">
-                          <input
-                              type="checkbox"
-                              checked
-                              readOnly
-                              className="form-check-input"
-                              onClick={() => completeTasks(t.name)}
-                          />
-                        </td>
-                        <td
-                            className="align-middle"
-                            style={{textDecoration: "line-through"}}
+                t.completed && (
+                  <tr key={t._id}>
+                    <td className="align-middle">
+                      <input
+                        type="checkbox"
+                        checked
+                        readOnly
+                        className="form-check-input"
+                        onClick={() => completeTasks(t.name)}
+                      />
+                    </td>
+                    <td
+                      className="align-middle"
+                      style={{ textDecoration: "line-through" }}
+                    >
+                      {t.name}
+                    </td>
+                    <td className="text-end align-middle">
+                      <Button
+                        onClick={(e) => deleteTasks(t.name, e)}
+                        className="delete"
+                        variant="danger"
+                      >
+                        <svg
+                          xmlns="http://www.w3.org/2000/svg"
+                          width="18"
+                          height="18"
+                          fill="currentColor"
+                          className="bi bi-trash"
+                          viewBox="0 0 16 16"
                         >
-                          {t.name}
-                        </td>
-                        <td className="text-end align-middle">
-                          <Button
-                              onClick={(e) => deleteTasks(t.name, e)}
-                              className="delete"
-                              variant="danger"
-                          >
-                            <svg
-                                xmlns="http://www.w3.org/2000/svg"
-                                width="18"
-                                height="18"
-                                fill="currentColor"
-                                className="bi bi-trash"
-                                viewBox="0 0 16 16"
-                            >
-                              <path
-                                  d="M5.5 5.5A.5.5 0 0 1 6 6v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5m2.5 0a.5.5 0 0 1 .5.5v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5m3 .5a.5.5 0 0 0-1 0v6a.5.5 0 0 0 1 0z"/>
-                              <path
-                                  d="M14.5 3a1 1 0 0 1-1 1H13v9a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V4h-.5a1 1 0 0 1-1-1V2a1 1 0 0 1 1-1H6a1 1 0 0 1 1-1h2a1 1 0 0 1 1 1h3.5a1 1 0 0 1 1 1zM4.118 4 4 4.059V13a1 1 0 0 0 1 1h6a1 1 0 0 0 1-1V4.059L11.882 4zM2.5 3h11V2h-11z"/>
-                            </svg>
-                          </Button>
-                        </td>
-                      </tr>
-                  ),
-          )}
+                          <path d="M5.5 5.5A.5.5 0 0 1 6 6v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5m2.5 0a.5.5 0 0 1 .5.5v6a.5.5 0 0 1-1 0V6a.5.5 0 0 1 .5-.5m3 .5a.5.5 0 0 0-1 0v6a.5.5 0 0 0 1 0z" />
+                          <path d="M14.5 3a1 1 0 0 1-1 1H13v9a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V4h-.5a1 1 0 0 1-1-1V2a1 1 0 0 1 1-1H6a1 1 0 0 1 1-1h2a1 1 0 0 1 1 1h3.5a1 1 0 0 1 1 1zM4.118 4 4 4.059V13a1 1 0 0 0 1 1h6a1 1 0 0 0 1-1V4.059L11.882 4zM2.5 3h11V2h-11z" />
+                        </svg>
+                      </Button>
+                    </td>
+                  </tr>
+                ),
+            )}
           </tbody>
         </Table>
       </Card.Body>
